+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Bump Google Services plugin to v4.3.12
 - Don't show no patients found error view when search query is empty
 - Don't show empty search history items in overdue search screen
+- Fix old synced appointments not getting deleted on purge
 
 ## 2022-06-20-8300